compare groups to threshold as discussed with @snigdhagit via email
Presently, a group is active if its norm multiplied by a tolerance
multiplied by the norm of the overall solution is nonzero. The only
way tolerance could have a consequence is if a floating point
arithmetic rounding error makes the norm so close to 0 it numerically
rounds to 0. This is presumably not the desired behavior, so it now
explicitly does a comparison