fix bug for atomic special case of group lasso
if groups are atomic, then the Jacobian is constant and shouldn't
contribute to the gradient. This is currently handled by setting the
values to be identically 0. However, this causes a broadcasting
problem, since the computation of the gradient involves pushing the
gradient of the jacobian through a bunch of matrices. If it's a
scalar 0, then numpy returns an *array* of 0's, and then everything
else gets broadcast up. We then get a matrix for our gradient rather
than a vector, and this makes the sampler fall down.

Because log_jacob[1] is sometimes an array (in the usual case) and
sometimes identically the scalar 0, I use the "is" boolean, which
seems hacky but I don't know what I should do instead