support passing XrawE through to posterior
If we do some sort of upstream "generalization" of the group
Lasso (e.g., standardized group lasso or overlapping groups), then the
design that the solver saw is different than the actual design. We
need a way to pass the selected variables in terms of the original
design to the posterior sampler so that it can adjust a variety of
quantities (chiefly through selected_targets).

This commit adds the XrawE argument to the posterior class's
init_method. By default, it is False. It passes this argument to the
selected_targets call. If XrawE is (identical to) False, then it runs
the way it used to. If not, it uses XrawE whenever it needs the
selected variables in terms of the original (raw) design matrix