be extra careful to avoid appearance of overwriting X
Now, W is the version of the data that is passed to the group lasso
object. If we are running in neither std or og mode, then W =
X.copy (the copy is to be *extra* careful that we are not mutating X
since we need its meaning to be preserved in outer scopes)

This also fixes a bug where in OG mode, it created the special matrix
W but then didn't do anything with it. Because W is now used when
constructing conv, the OG-constructed W will get passed through.