tweak data generation to work in OG mode
if we have overlapping groups, we specify our groups as a dictionary,
but the gaussian_group_instance fn expects an array assigning each
predictor to one group. It *only* uses grouping information when
setting beta, so all that really matters is which features are
active (in at least one active group) and which are inactive (in *no*
active groups). Loop over the first traj.sgroup many groups and
collect the indices, gather all of these predictors into group 0, set
the rest to group 1, and set sgroup so that the first (0) group is
active and the second (1) is inactive