have grp_lasso_selection translate from nonzero to nonzero_raw
When grp_lasso_selection is called with overlapping groups, it will do
data duplication and then fit a model to W, and its nonzero entries
will correspond to this. This commit tweaks it so that it will
translate these nonzero entries back into the original indices, which
is all that the "query" really cares about. When fitting the posi
method, we need to give it XrawE (the original design matrix) for the
posterior, so we adjust that, too