hardcode OG group assignments
pypet won't let us set a parameter to a dict, or nested list, or
anything else sufficiently exotic for us to specify overlapping
groups. There may be an elegant workaround, but instead this commit
will "hardcode" our overlapping scheme into the experiment code. If
the "og" parameter is set to true, it will create the group dict on
the fly and use that instead. It's ugly, but it works...