scale weights (lambda) by the size of the training set
currently, our loss is not normalized by n. we can fix this by scaling
lambda appropriately. We compute n_scale, which is the fraction of the
data used for the query out of all available data, and then scale each
weight (lambda_g) by this quantity. When we query the whole data (as
in naive or posi) this will be 1 and have no consequence. If we do
e.g., 50/50 data splitting, this will halve the contribution of the
penalty (which should correspond to effectively halving the loss that
occurred due to using only half the data for the query)